api: move mesh/tools.ts → tools/mesh.ts for consistency
All agent-callable tool factories should live under packages/api/src/tools/.
The mesh/ subdir keeps only the broker infrastructure (MeshServer +
wire types).

Before:
  mesh/{server,tools,types}.ts
  tools/hierarchy.ts (etc)

After:
  mesh/{server,types}.ts
  tools/{hierarchy,mesh,...}.ts

Symmetric with hierarchy/work-product/memory tools — agent-callable
tools all live together, broker classes live in their own subsystem.
